requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.13.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
inflection==0.5.1
openpyxl==3.1.5
packaging==25.0
pillow==12.1.0
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
shortuuid==1.0.13
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
